Refactor example a bunch. Fix data-lake and some refactors to toS3
