+++ packages/punchcard/lib/enumerable/index.ts
@@ -2,7 +2,7 @@ export * from './collector';
 export * from './enumerable';
 export * from './glue';
 export * from './queue';
-export * from './s3';
+export * from './s3-delivery-stream';
 export * from './sink';
 export * from './stream';
 export * from './topic';
